+++ R/find_formula.R
@@ -279,6 +279,16 @@ find_formula.maxim <- find_formula.default
 
 
 #' @export
+find_formula.selection <- function(x, ...) {
+  model_call <- parse(text = deparse(get_call(m)))[[1]]
+  f <- list(conditional = list(selection = stats::as.formula(model_call$selection),
+                               outcome = stats::as.formula(model_call$outcome)))
+  attr(f, "two_stage") <- TRUE
+  .find_formula_return(f)
+}
+
+
+#' @export
 find_formula.svy_vglm <- function(x, ...) {
   find_formula(x$fit)
 }