+++ R/find_response.R
@@ -55,6 +55,14 @@ find_response.model_fit <- function(x, combine = TRUE, ...) {
 
 
 #' @export
+find_response.selection <- function(x, combine = TRUE, ...) {
+  f <- find_formula(x)
+  resp <- c(.safe_deparse(f$conditional$selection[[2L]]), .safe_deparse(f$conditional$outcome[[2L]]))
+  check_cbind(resp, combine, model = x)
+}
+
+
+#' @export
 find_response.mediate <- function(x, combine = TRUE, ...) {
   f <- find_formula(x)