+++ R/get_parameters.R
@@ -125,6 +125,28 @@ get_parameters.model_fit <- function(x, ...) {
 
 
 #' @export
+get_parameters.selection <- function(x, component = c("all", "selection", "outcome", "auxiliary"), ...) {
+  component <- match.arg(component)
+  s <- summary(x)
+  params <- data.frame(
+    Parameter = row.names(s$estimate),
+    Estimate = s$estimate[[1]],
+    Component = "auxiliary",
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+  params$Component[s$param$index$betaS] <- "selection"
+  params$Component[s$param$index$betaO] <- "outcome"
+
+  if (component != "all") {
+    params <- params[params$Component == component, , drop = FALSE]
+  }
+
+  .remove_backticks_from_parameter_names(params)
+}
+
+
+#' @export
 get_parameters.epi.2by2 <- function(x, ...) {
   coef_names <- grepl(".strata.wald", names(x$massoc), fixed = TRUE)
   cf <- x$massoc[coef_names]