+++ R/is_model.R
@@ -119,9 +119,9 @@ is_regression_model <- function(x) {
     "rjags", "rlm", "rlme", "rlmerMod", "RM", "rma", "rmanovab",
     "rma.uni", "rms", "robmixglm", "robtab", "rq", "rqs", "rqss",
     # s --------------------
-    "sarlm", "scam", "sem", "SemiParBIV", "slm", "speedlm", "speedglm", "splmm",
-    "spml", "stanmvreg", "stanreg", "summary.lm", "survfit", "survreg",
-    "survPresmooth", "svyglm", "svy_vglm", "svyolr",
+    "sarlm", "scam", "selection", "sem", "SemiParBIV", "slm", "speedlm",
+    "speedglm", "splmm", "spml", "stanmvreg", "stanreg", "summary.lm",
+    "survfit", "survreg", "survPresmooth", "svyglm", "svy_vglm", "svyolr",
     # t --------------------
     "t1way", "t2way", "t3way", "test_mediation", "tobit", "trendPMCMR",
     "truncreg",